Refactor: MockBean으로 인한 테스트 실패로 '액세스 토큰과 일치하는 사용자 체크' 테스트를 따로 분리
